Started Work on B2 Model
For the B2 model I had two approaches in mind:
  - Use dlib method to extract 68 featrues then thr features around
  eyes crop the image to just one particular eye.
  - Use opencv Hough Circles to approximate the eye as circle then
  crop the image around.

Both approaches have their pitfalls in mind. The dlib method requires
a larger amount of processing power but could provide a better accuracy.
Similary, opencv will be faster but may not be as accurate. I have
decided to go with opencv approach for not only the faster performance
but this approach allows me to easily "ignore/reject" images where the
person is wearing glasses