Copied over Base Model Class
The main method to implement is convert_to_feature which takes
an image path and converts it into data to be used by sklearn.
The base class otherwise is same as the one from task A

As stated in earlier commit messages there are few specialised
approaches I have in mind that I plan to try on to convert this
data. As, well as a basic approach reading, resizing, removing
noise from the image and testing it out.

In the next commit I'll try out the basic approach first and test
if we can get decent accuracy with it.