Implemented the OpenCV Hough Circle Feature Extraction
I was able to get roughly 85 to 86% accuracy. However, a major
problem I noticed was number of images getting rejected because
character were wearing "glasses" thus no circles could be detected.

I tried to do test out various parameters, and saw accuracy and number
of images processed were inversely related, which is to be expected. I
was able to increase the accuracy of the model to be 90% but in turn
over 3000 images were getting rejected. The current chosen threshold
parameters give us the best ratio of accuracy to number of images.